@@ -1,27 +1,25 @@
-import { useEffect } from "react";
-import { useHistory } from "react-router-dom";
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
 
-import QuoteForm from "../components/quotes/QuoteForm";
-import useHttp from "../hooks/hooks/use-http";
-import { addQuote } from "../lib/lib/api";
+import QuoteForm from '../components/quotes/QuoteForm';
+import useHttp from '../hooks/use-http';
+import { addQuote } from '../lib/api';
 
 const NewQuote = () => {
   const { sendRequest, status } = useHttp(addQuote);
-  const history = useHistory();
+  const navigate = useNavigate();
 
   useEffect(() => {
-    if (status === "completed") {
-      history.push("/quotes");
+    if (status === 'completed') {
+      navigate('/quotes');
     }
-  }, [status, history]);
+  }, [status, navigate]);
 
   const addQuoteHandler = (quoteData) => {
     sendRequest(quoteData);
   };
 
-  return (
-    <QuoteForm isLoading={status === "pending"} onAddQuote={addQuoteHandler} />
-  );
+  return <QuoteForm isLoading={status === 'pending'} onAddQuote={addQuoteHandler} />;
 };
 
 export default NewQuote;
