@@ -1,9 +1,9 @@
-import { useRef, useEffect } from "react";
+import { useRef, useEffect } from 'react';
 
-import useHttp from "../../hooks/hooks/use-http";
-import { addComment } from "../../lib/lib/api";
-import LoadingSpinner from "../UI/LoadingSpinner";
-import classes from "./NewCommentForm.module.css";
+import useHttp from '../../hooks/use-http';
+import { addComment } from '../../lib/api';
+import LoadingSpinner from '../UI/LoadingSpinner';
+import classes from './NewCommentForm.module.css';
 
 const NewCommentForm = (props) => {
   const commentTextRef = useRef();
@@ -13,7 +13,7 @@ const NewCommentForm = (props) => {
   const { onAddedComment } = props;
 
   useEffect(() => {
-    if (status === "completed" && !error) {
+    if (status === 'completed' && !error) {
       onAddedComment();
     }
   }, [status, error, onAddedComment]);
@@ -30,17 +30,17 @@ const NewCommentForm = (props) => {
 
   return (
     <form className={classes.form} onSubmit={submitFormHandler}>
-      {status === "pending" && (
-        <div className="centered">
+      {status === 'pending' && (
+        <div className='centered'>
           <LoadingSpinner />
         </div>
       )}
       <div className={classes.control} onSubmit={submitFormHandler}>
-        <label htmlFor="comment">Your Comment</label>
-        <textarea id="comment" rows="5" ref={commentTextRef}></textarea>
+        <label htmlFor='comment'>Your Comment</label>
+        <textarea id='comment' rows='5' ref={commentTextRef}></textarea>
       </div>
       <div className={classes.actions}>
-        <button className="btn">Add Comment</button>
+        <button className='btn'>Add Comment</button>
       </div>
     </form>
   );
