@@ -9,12 +9,12 @@ const MainNavigation = () => {
       <nav className={classes.nav}>
         <ul>
           <li>
-            <NavLink to='/quotes' activeClassName={classes.active}>
+            <NavLink to='/quotes' className={navData => navData.isActive ? classes.active : '' }>
               All Quotes
             </NavLink>
           </li>
           <li>
-            <NavLink to='/new-quote' activeClassName={classes.active}>
+            <NavLink to='/new-quote' className={navData => navData.isActive ? classes.active : '' }>
               Add a Quote
             </NavLink>
           </li>
