@@ -3,8 +3,8 @@ import { useEffect } from "react";
 import QuoteList from "../components/quotes/QuoteList";
 import LoadingSpinner from "../components/UI/LoadingSpinner";
 import NoQuotesFound from "../components/quotes/NoQuotesFound";
-import useHttp from "../hooks/hooks/use-http";
-import { getAllQuotes } from "../lib/lib/api";
+import useHttp from "../hooks/use-http";
+import { getAllQuotes } from "../lib/api";
 
 const AllQuotes = () => {
   const {
