@@ -1,39 +1,35 @@
-import { Fragment, useEffect } from "react";
-import { useParams, Route, Link, useRouteMatch } from "react-router-dom";
+import { Fragment, useEffect } from 'react';
+import { useParams, Outlet } from 'react-router-dom';
 
-import HighlightedQuote from "../components/quotes/HighlightedQuote";
-import Comments from "../components/comments/Comments";
-import useHttp from "../hooks/hooks/use-http";
-import { getSingleQuote } from "../lib/lib/api";
-import LoadingSpinner from "../components/UI/LoadingSpinner";
+import HighlightedQuote from '../components/quotes/HighlightedQuote';
+import useHttp from '../hooks/use-http';
+import { getSingleQuote } from '../lib/api';
+import LoadingSpinner from '../components/UI/LoadingSpinner';
 
 const QuoteDetail = () => {
-  const match = useRouteMatch();
   const params = useParams();
 
   const { quoteId } = params;
 
-  const {
-    sendRequest,
-    status,
-    data: loadedQuote,
-    error,
-  } = useHttp(getSingleQuote, true);
+  const { sendRequest, status, data: loadedQuote, error } = useHttp(
+    getSingleQuote,
+    true
+  );
 
   useEffect(() => {
     sendRequest(quoteId);
   }, [sendRequest, quoteId]);
 
-  if (status === "pending") {
+  if (status === 'pending') {
     return (
-      <div className="centered">
+      <div className='centered'>
         <LoadingSpinner />
       </div>
     );
   }
 
   if (error) {
-    return <p className="centered">{error}</p>;
+    return <p className='centered'>{error}</p>;
   }
 
   if (!loadedQuote.text) {
@@ -43,16 +39,7 @@ const QuoteDetail = () => {
   return (
     <Fragment>
       <HighlightedQuote text={loadedQuote.text} author={loadedQuote.author} />
-      <Route path={match.path} exact>
-        <div className="centered">
-          <Link className="btn--flat" to={`${match.url}/comments`}>
-            Load Comments
-          </Link>
-        </div>
-      </Route>
-      <Route path={`${match.path}/comments`}>
-        <Comments />
-      </Route>
+      <Outlet />
     </Fragment>
   );
 };
